WIP: splitting cluster-aware tests away from scalding-core
