+++ project/plugins.sbt
@@ -19,3 +19,5 @@ addSbtPlugin("org.scoverage" % "sbt-scoverage" % "1.0.4")
 addSbtPlugin("org.xerial.sbt" % "sbt-sonatype" % "1.0")
 addSbtPlugin("pl.project13.scala" % "sbt-jmh" % "0.2.2")
 addSbtPlugin("org.brianmckenna" % "sbt-wartremover" % "0.14")
+
+addSbtPlugin("io.get-coursier" % "sbt-coursier" % "1.0.0-M14")
