+++ scalding-core-fabric-tests/src/fabric/scala/com/twitter/scalding/ExecutionTest.scala
