+++ scalding-core-fabric-tests/src/fabric/scala/com/twitter/scalding/ReduceOperationsTest.scala
