+++ scalding-core-fabric-tests/src/fabric/scala/com/twitter/scalding/TypedPipeTest.scala
@@ -16,8 +16,9 @@ limitations under the License.
 package com.twitter.scalding
 
 import org.scalatest.{ FunSuite, Matchers, WordSpec }
+import java.lang.SuppressWarnings
 
-import com.twitter.scalding.source.TypedText
+import com.twitter.scalding.source.{ TypedSep, TypedText }
 // Use the scalacheck generators
 import org.scalacheck.Gen
 import scala.collection.mutable.Buffer
@@ -847,8 +848,8 @@ trait TypedComplexHashAndMergeJobBase {
   def tdXe: TypedPipe[String]
 }
 
-class TypedComplexHashAndMergeJob(args: Args ,
-                                  fieldsToMerge: Seq[(String, TypedComplexHashAndMergeJobBase => TypedPipe[String])])
+class TypedComplexHashAndMergeJob(args: Args,
+  fieldsToMerge: Seq[(String, TypedComplexHashAndMergeJobBase => TypedPipe[String])])
   extends Job(args) with TypedComplexHashAndMergeJobBase {
 
   override def name: String = super.name + " (" + fieldsToMerge.map(_._1).mkString(" ++ ") + ")"
@@ -874,7 +875,7 @@ class TypedComplexHashAndMergeTest extends WordSpec with Matchers {
     ("a", _.ta),
     ("a∩b", _.taXb),
     ("c", _.tc),
-    ("d",  _.td),
+    ("d", _.td),
     ("d∩e", _.tdXe))
 
   val selection = fields.permutations.take(3) // Take'em all if you need to prove all permutations work equally (kind of slow, and internally we do use commutativity)
@@ -931,8 +932,8 @@ class TypedTwistedHashAndMergeJob(args: Args) extends Job(args) {
     (y.head, y.tail)
   }).group
 
-  val twistAB: TypedPipe[String] = twistA.hashJoin(twistB).values.map { case (a,b) => a + "≡" + b }
-  val twistBA: TypedPipe[String] = twistB.hashJoin(twistA).values.map { case (a,b) => a + "≢" + b }
+  val twistAB: TypedPipe[String] = twistA.hashJoin(twistB).values.map { case (a, b) => a + "≡" + b }
+  val twistBA: TypedPipe[String] = twistB.hashJoin(twistA).values.map { case (a, b) => a + "≢" + b }
 
   (taXbXe ++ tdXeXb ++ twistAB ++ twistBA)
     .write(TypedText.tsv[String]("output"))
@@ -941,7 +942,7 @@ class TypedTwistedHashAndMergeJob(args: Args) extends Job(args) {
 class TypedTwistedHashAndMergeTest extends WordSpec with Matchers {
   import Dsl._
 
-  s"A TypedTwistedHashAndMergeTest" should {
+  "A TypedTwistedHashAndMergeTest" should {
     var idx = 0
 
     JobTest(new TypedTwistedHashAndMergeJob(_: Args))
