+++ scalding-core-fabric-tests/src/fabric/scala/com/twitter/scalding/mathematics/Matrix2Test.scala
