--- scalding-core-fabric-tests/src/main/scala/com/twitter/scalding/FabricTests.scala
@@ -1,5 +1,0 @@
-package com.twitter.scalding
-
-import org.scalatest.Suites
-
-class FabricTests extends Suites(new FileSourceHadoopExtraTest, new typed.TypedFabricTests)
