--- scalding-core-fabric-tests/src/main/scala/com/twitter/scalding/typed/TypedFabricTests.scala
@@ -1,6 +1,0 @@
-package com.twitter.scalding.typed
-
-import org.scalatest.Suites
-
-class TypedFabricTests extends Suites(new NoStackLineNumberTest, new PartitionedDelimitedTest, new PartitionedTextLineTest)
-
