--- scalding-fabric-flink/src/test/scala/com/twitter/scalding/FlinkFabricTest.scala
@@ -1,6 +1,0 @@
-package com.twitter.scalding
-
-import org.scalatest.Suites
-
-class FlinkFabricTest extends Suites(new FabricTests)
-
