--- scalding-fabric-hadoop/src/test/scala/com/twitter/scalding/LegacyHadoopFabricTest.scala
@@ -1,6 +1,0 @@
-package com.twitter.scalding
-
-import org.scalatest.Suites
-
-class LegacyHadoopFabricTest extends Suites(new FabricTests)
-
