--- scalding-fabric-hadoop2-mr1/src/test/scala/com/twitter/scalding/Hadoop2Mr1FabricTest.scala
@@ -1,6 +1,0 @@
-package com.twitter.scalding
-
-import org.scalatest.Suites
-
-class Hadoop2Mr1FabricTest extends Suites(new FabricTests)
-
