--- scalding-fabric-tez/src/main/scala/com/twitter/scalding/TezFabricTest.scala
@@ -1,6 +1,0 @@
-package com.twitter.scalding
-
-import org.scalatest.Suites
-
-class TezFabricTest extends Suites(new FabricTests)
-
